requirements.txt
streamli
openai
google
generativeai